outside flattening working
