+++ Sources/Provider/CollectionComposer.swift
@@ -8,7 +8,9 @@
 
 import UIKit
 
-open class CollectionComposer: BaseCollectionProvider {
+open class CollectionComposer: ComposedCollectionProvider, CollectionReloadable {
+
+  public var identifier: String?
 
   public var sections: [AnyCollectionProvider] {
     didSet { setNeedsReload() }
@@ -29,80 +31,58 @@ open class CollectionComposer: BaseCollectionProvider {
     self.presenter = presenter
     self.layout = layout
     self.sections = sections
-    super.init(identifier: identifier)
+    self.identifier = identifier
   }
 
   public convenience init(identifier: String? = nil,
                           layout: CollectionLayout = FlowLayout(),
                           presenter: CollectionPresenter? = nil,
                           _ sections: AnyCollectionProvider...) {
-    self.init(layout: layout, presenter: presenter, sections: sections)
+    self.init(identifier: identifier, layout: layout, presenter: presenter, sections: sections)
   }
 
-  open override var numberOfItems: Int {
+  open var numberOfItems: Int {
     return sections.count
   }
 
-  open override var hasSection: Bool {
-    return true
-  }
-
-  open override func section(at: Int) -> AnyCollectionProvider? {
+  open func section(at: Int) -> AnyCollectionProvider? {
     return sections[at]
   }
 
-  open override func view(at: Int) -> UIView {
-    fatalError()
-  }
-
-  open override func update(view: UIView, at: Int) {
-    fatalError()
-  }
-
-  open override func identifier(at: Int) -> String {
+  open func identifier(at: Int) -> String {
     return sections[at].identifier ?? "\(at)"
   }
 
-  open override func layout(collectionSize: CGSize) {
+  open func layout(collectionSize: CGSize) {
     layout.layout(context: CollectionComposerLayoutContext(collectionSize: collectionSize,
                                                            sections: sections))
   }
 
-  open override var contentSize: CGSize {
+  open var contentSize: CGSize {
     return layout.contentSize
   }
 
-  open override func visibleIndexes(visibleFrame: CGRect) -> [Int] {
+  open func visibleIndexes(visibleFrame: CGRect) -> [Int] {
     return layout.visibleIndexes(visibleFrame: visibleFrame)
   }
 
-  open override func frame(at: Int) -> CGRect {
+  open func frame(at: Int) -> CGRect {
     return layout.frame(at: at)
   }
 
-  open override func presenter(at: Int) -> CollectionPresenter? {
-    fatalError()
-  }
-
-  open override func willReload() {
-    super.willReload()
+  open func willReload() {
     for section in sections {
       section.willReload()
     }
   }
 
-  open override func didReload() {
-    super.didReload()
+  open func didReload() {
     for section in sections {
       section.didReload()
     }
   }
 
-  open override func didTap(view: UIView, at: Int) {
-    fatalError()
-  }
-
-  open override func hasReloadable(_ reloadable: CollectionReloadable) -> Bool {
+  open func hasReloadable(_ reloadable: CollectionReloadable) -> Bool {
     return reloadable === self || sections.contains(where: { $0.hasReloadable(reloadable) })
   }
 }
